+++ LOLInfoApp/LOLInfoApp/Scenes/Champion/Main/PresentationLayer/CollectionView/ChampionMainCollectionViewScetionType.swift
@@ -1,0 +1,13 @@
+//
+//  ChampionMainCollectionViewSectionType.swift
+//  LOLInfoApp
+//
+//  Created by Jeong Deokho on 2024/03/29.
+//
+
+import Foundation
+
+enum ChampionMainCollectionViewSectionType: Int {
+    case favoriteList
+    case mainList
+}
